refactor: rename input properties to avoid collisions with dom
BREAKING CHANGE: all of the input events have been renamed to avoid
collisions and errors with dom. All names containing Mouse are now named
MouseData (to avoid collisions with MouseEvent and others) and to make
things consistent, all Keypress and Keyboard are named KeyData. This is
because to generate these events, we listen to the `data` event on the
`stdin` and `Key` is shorter than `Keyboard`.

- `*Keypress*` -> `*KeyData*`
- `*Keyboard*` -> `*KeyData*`
- `*Mouse*` -> `*MouseData*`